remove old test generator.
